Base for edit medications page
