+++ app/src/main/java/com/example/dosediary/view/EditMedication.kt
@@ -1,16 +1,124 @@
-package com.example.dosediary.view
+package com.example.DoseDiary
 
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.TopAppBar
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
+import android.app.DatePickerDialog
+import android.app.TimePickerDialog
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.material3.*
+import androidx.compose.runtime.*
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.text.input.KeyboardType
+import androidx.compose.ui.unit.dp
+import java.util.*
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun EditMedication() {
-    Scaffold(
-        topBar = { TopAppBar(title = { Text("Edit Medication") }) },
-        content = { }
-    )
+    val medicationName = remember { mutableStateOf("Ibuprofen") }
+    val effectivenessOptions = listOf("Effective", "Moderate", "Marginal", "Ineffective")
+    val selectedEffectiveness = remember { mutableStateOf(effectivenessOptions[0]) }
+
+    val calendar = Calendar.getInstance()
+    calendar.set(2023, Calendar.MAY, 10, 10, 30)
+    val date = remember { mutableStateOf(calendar.time) }
+    val time = remember { mutableStateOf(calendar.time) }
+
+    val context = LocalContext.current
+
+    Column(
+        modifier = Modifier
+            .fillMaxSize()
+            .padding(16.dp)
+    ) {
+        Text(
+            text = "Edit Medication History",
+            style = MaterialTheme.typography.titleLarge,
+            modifier = Modifier.padding(bottom = 16.dp)
+        )
+
+        OutlinedTextField(
+            value = medicationName.value,
+            onValueChange = { medicationName.value = it },
+            label = { Text("Medication Name") },
+            modifier = Modifier.fillMaxWidth()
+        )
+
+        Spacer(modifier = Modifier.height(16.dp))
+
+        ExposedDropdownMenuBox(
+            expanded = false,
+            onExpandedChange = { }
+        ) {
+            OutlinedTextField(
+                value = selectedEffectiveness.value,
+                onValueChange = { },
+                readOnly = true,
+                label = { Text("Effectiveness") },
+                trailingIcon = {
+                    ExposedDropdownMenuDefaults.TrailingIcon(expanded = false)
+                },
+                modifier = Modifier.fillMaxWidth()
+            )
+            ExposedDropdownMenu(
+                expanded = false,
+                onDismissRequest = { }
+            ) {
+                effectivenessOptions.forEach { effectiveness ->
+                    DropdownMenuItem(
+                        text = { Text(effectiveness) },
+                        onClick = {
+                            selectedEffectiveness.value = effectiveness
+                        }
+                    )
+                }
+            }
+        }
+
+        Spacer(modifier = Modifier.height(16.dp))
+
+        val datePickerDialog = DatePickerDialog(
+            context,
+            { _, year, month, dayOfMonth ->
+                calendar.set(year, month, dayOfMonth)
+                date.value = calendar.time
+            },
+            calendar.get(Calendar.YEAR),
+            calendar.get(Calendar.MONTH),
+            calendar.get(Calendar.DAY_OF_MONTH)
+        )
+
+        val timePickerDialog = TimePickerDialog(
+            context,
+            { _, hourOfDay, minute ->
+                calendar.set(Calendar.HOUR_OF_DAY, hourOfDay)
+                calendar.set(Calendar.MINUTE, minute)
+                time.value = calendar.time
+            },
+            calendar.get(Calendar.HOUR_OF_DAY),
+            calendar.get(Calendar.MINUTE),
+            false
+        )
+
+        OutlinedTextField(
+            value = remember { mutableStateOf("${calendar.get(Calendar.MONTH) + 1}/${calendar.get(Calendar.DAY_OF_MONTH)}/${calendar.get(Calendar.YEAR)}") }.value,
+            onValueChange = { },
+            label = { Text("Select Date") },
+            readOnly = false,
+            modifier = Modifier.fillMaxWidth().clickable { datePickerDialog.show() },
+            keyboardOptions = KeyboardOptions.Default.copy(keyboardType = KeyboardType.Number)
+        )
+
+        Spacer(modifier = Modifier.height(16.dp))
+
+        OutlinedTextField(
+            value = remember { mutableStateOf("${calendar.get(Calendar.HOUR_OF_DAY)}:${calendar.get(Calendar.MINUTE)} ${if (calendar.get(Calendar.AM_PM) == Calendar.AM) "AM" else "PM"}") }.value,
+            onValueChange = { },
+            label = { Text("Select Time") },
+            readOnly = false,
+            modifier = Modifier.fillMaxWidth().clickable { timePickerDialog.show() },
+            keyboardOptions = KeyboardOptions.Default.copy(keyboardType = KeyboardType.Number)
+        )
+    }
 }
